Documentation Implementation

+++ src/function/support/s21_lazy_downsize.c
@@ -15,8 +15,8 @@ uint8_t s21_lazy_downsize(s21_decimal_lazy *value_1,
   uint16_t new_size = 0;
   s21_decimal_lazy *ptr = NULL;
 
-  ptr = (s21_decimal_lazy *)((value_1->size > value_2->size) * (long)value_1 +
-                             (value_1->size < value_2->size) * (long)value_2);
+  ptr = (s21_decimal_lazy *)((value_1->size > value_2->size) * (uintptr_t)value_1 +
+                             (value_1->size < value_2->size) * (uintptr_t)value_2);
 
   new_size = (value_1->size > value_2->size) * value_2->size +
              (value_1->size <= value_2->size) * value_1->size;
